Small fixes to real management files

+++ src/real_kernels.cu
@@ -5,7 +5,7 @@
 
 #if REAL == HALF
 
-__global__ void float2real_array_kernel(float* src, real_device* dst, size_t n) {
+__global__ void float2real_array_kernel(float* src, real_device* dst, int n) {
 	int i = (blockIdx.x + blockIdx.y * gridDim.x) * blockDim.x + threadIdx.x;
 	if (i < n) {
 		#if REAL == HALF
@@ -16,12 +16,12 @@ __global__ void float2real_array_kernel(float* src, real_device* dst, size_t n) 
 	}
 }
 
-void float2real_array_gpu(float* src, real* dst, size_t n) {
+void float2real_array_gpu(float* src, real* dst, int n) {
 	float2real_array_kernel<<<cuda_gridsize(n), BLOCK>>>(src, (real_device*)dst, n);
 	check_error(cudaPeekAtLastError());
 }
 
-__global__ void real2float_array_kernel(real_device* src, float* dst, size_t n) {
+__global__ void real2float_array_kernel(real_device* src, float* dst, int n) {
 	int i = (blockIdx.x + blockIdx.y * gridDim.x) * blockDim.x + threadIdx.x;
 	if (i < n) {
 		#if REAL == HALF
@@ -32,7 +32,7 @@ __global__ void real2float_array_kernel(real_device* src, float* dst, size_t n) 
 	}	
 }
 
-void real2float_array_gpu(real* src, float* dst, size_t n) {
+void real2float_array_gpu(real* src, float* dst, int n) {
 	real2float_array_kernel<<<cuda_gridsize(n), BLOCK>>>((real_device*)src, dst, n);
 	check_error(cudaPeekAtLastError());
 }
